Add support for reading system compressed files
System compression is a feature that allows rarely modified files to be
compressed using the XPRESS or LZX compression formats.  It is not built
directly into NTFS but rather is implemented using reparse points.  This
feature appeared in Windows 10 and it appears that many Windows 10
systems have been using it by default.

This patch adds read support only; write support is probably best
delegated to a third-party tool.